+++ bin/domain/usecases/normalizeReportBody.js
@@ -1,0 +1,6 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+exports.normalizeReportBody = void 0;
+const normalizeReportBody = (body) => body.replace(/\\`/g, '`');
+exports.normalizeReportBody = normalizeReportBody;
+//# sourceMappingURL=normalizeReportBody.js.map